Move remaining UI4 components -> /common
