+++ src/components/common/BlurBackground.tsx
